Realtime_Tweet_Analyser/requirements.txt
tweepy
kafka-python
hazm
elasticsearch
cassandra-driver
redis
Flask
flask_paginate
clickhouse-driver